+++ src/components/CommitTextField/index.ts
@@ -1,2 +1,0 @@
-import CommitTextField from "./component";
-export default CommitTextField;